Extract FilepickerImageUrl class into separate file and RDoc documentation
