+++ app/helpers/filepicker_rails/application_helper.rb
@@ -172,80 +172,6 @@ module FilepickerRails
       FilepickerImageUrl.new(url, options).execute
     end
 
-    class FilepickerImageUrl
-
-      CONVERT_OPTIONS = [:w, :h, :fit, :align, :rotate, :crop, :format,
-                         :quality, :watermark, :watersize, :waterposition,
-                         :crop_first]
-      VALID_OPTIONS   = CONVERT_OPTIONS + [:cache]
-
-      def initialize(url, options = {})
-        @url, @options = url, options
-      end
-
-      def execute
-        base_url = url_with_path.split('?').first
-        query_params = all_options.to_query
-
-        [base_url, query_params.presence].compact.join('?')
-      end
-
-      private
-
-        attr_reader :url, :options
-
-        def valid_options
-          options.select { |option| VALID_OPTIONS.include?(option) }
-        end
-
-        def convert_options
-          options.select { |option| CONVERT_OPTIONS.include?(option) }
-        end
-
-        def all_options
-          [original_url_options, valid_options, policy_config].inject(&:merge)
-        end
-
-        def original_url_options
-          query_string = url_with_path.split('?')[1]
-
-          if query_string
-            Rack::Utils.parse_nested_query(query_string)
-          else
-            {}
-          end
-        end
-
-        def cdn_host
-          @cdn_host ||= ::Rails.application.config.filepicker_rails.cdn_host
-        end
-
-        def cdn_url
-          if cdn_host
-            uri = URI.parse(url)
-            url.gsub("#{uri.scheme}://#{uri.host}", cdn_host)
-          else
-            url
-          end
-        end
-
-        def policy_config
-          Policy.apply
-        end
-
-        def url_with_path
-          @url_with_path ||= if append_convert_on_url_path?
-            "#{cdn_url}/convert"
-          else
-            cdn_url
-          end
-        end
-
-        def append_convert_on_url_path?
-          convert_options.any? && !cdn_url.match('/convert')
-        end
-    end
-
     private
 
       def export_widget(text, url, mimetype, options, &block)
