+++ lib/filepicker-rails.rb
@@ -1,5 +1,6 @@
 require "filepicker_rails/configuration"
 require "filepicker_rails/policy"
+require "filepicker_rails/filepicker_image_url"
 require "filepicker_rails/engine"
 require "filepicker_rails/tag"
